practice/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "guitar-pro"
version = "0.1.0"
description = "专业吉他练习应用 - 基于先进 MIR 技术的音频-乐谱对齐系统"
requires-python = ">=3.10"
dependencies = [
    "sounddevice>=0.4.6",
    "numpy>=1.24",
    "PySide6>=6.6",
    "pyqtgraph>=0.13",
]

[project.optional-dependencies]
mir = [
    "librosa>=0.10",
    "basic-pitch>=0.3",
    "crepe>=0.0.16",
    "madmom>=0.16",
    "pretty-midi>=0.2",
    "noisereduce>=3.0",
]
ml = [
    "onnxruntime-gpu>=1.17",
]
dev = [
    "pytest>=7.0",
    "black",
    "ruff",
]

[project.scripts]
guitar-pro = "src.main:main"
